refactor : 트랜잭션 적용
save 관련 문제가 있어 도서 리스트 저장시 트랜잭션을 설정해 이를 해결했습니다.
