+++ movie-book/src/main/java/hello/moviebook/book/service/BookService.java
@@ -37,7 +37,7 @@ public class BookService {
     private final UserRepository userRepository;
     private final UserLikeGenreRepository userLikeGenreRepository;
     private final UserDislikeGenreRepository userDislikeGenreRepository;
-    private final UserBookRepository userBookRepository;
+    private final UserBookService userBookService;
     private final WebClient.Builder webClientBuilder;
     private WebClient webClient;
 
@@ -170,7 +170,7 @@ public class BookService {
         if (bookNameList == null)
             return null;
 
-        saveUserBookList(user, bookNameList);
+        userBookService.saveUserBookList(user, bookNameList);
 
         return bookNameList;
     }
@@ -191,47 +191,5 @@ public class BookService {
         log.info("nbooks : {}", flaskReqDTO.getNBooks());
 
         return requestBody;
-    }
-
-    private void saveUserBookList(User user, List<String> recBookList) {
-        // 사용자와 관련된 기존 UserBook 데이터 조회
-        List<Book> existingBooks = userBookRepository.findAllByUser(user).stream()
-                .map(UserBook::getBook)
-                .toList();
-
-        // 추천 도서 중 기존 데이터에 없는 책 필터링
-        List<UserBook> userBookList = recBookList.stream()
-                .map(bookRepository::findBookByBookName)
-                .filter(book -> !existingBooks.contains(book)) // 기존 데이터에 없는 책만 필터링
-                .map(book -> UserBook.builder()
-                        .user(user)
-                        .book(book)
-                        .build())
-                .toList();
-
-        userBookRepository.saveAll(userBookList);
-    }
-
-    public List<BookDTO> getUserBookList(User user) {
-        List<UserBook> userBookList = userBookRepository.findAllByUser(user);
-
-        return userBookList.stream()
-                .map(userBook -> BookDTO.builder()
-                        .userBook(userBook)
-                        .build())
-                .toList();
-    }
-
-    public boolean ratingBook(User user, RatingBookReq ratingBookReq) {
-        UserBook userBook = userBookRepository.findByUserAndBook(user, bookRepository.findBookByIsbn(ratingBookReq.getIsbn()));
-
-        if (userBook == null)
-            return false;
-
-        userBook.setBookRating(ratingBookReq.getRating());
-        userBook.setBookReview(ratingBookReq.getReview());
-
-        userBookRepository.save(userBook);
-        return true;
     }
 }
